+++ src/pages/_app.tsx
@@ -1,3 +1,6 @@
+/* eslint-disable global-require */
+/* eslint-disable import/no-extraneous-dependencies */
+
 import '../styles/global.css';
 import '../styles/wordpress.css';
 
@@ -5,6 +8,14 @@ import { ApolloProvider } from '@apollo/client';
 import type { AppProps } from 'next/app';
 
 import { useApollo } from '@/lib/apolloClient';
+
+if (process.env.NODE_ENV === 'development') {
+  import('@impulse.dev/runtime').then((impulse) =>
+    impulse.run({
+      tailwindConfig: require('../../tailwind.config'),
+    })
+  );
+}
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   const apolloClient = useApollo(pageProps);